# api/openapi.yaml
openapi: 3.0.3
info:
  title: Musubi API
  description: High-Performance Network Scenario & SNMP Orchestration Engine with DDD Layered Architecture.
  version: 1.0.0
  contact:
    name: Musubi Core Team
servers:
  - url: http://localhost:8080
    description: Local / Air-Gapped Single Binary Server
...
paths:
...
  /v1/targets:
    get:
      summary: List Targets
      description: Retrieve registered network targets with optional filtering.
      operationId: listTargets
      tags:
        - Targets
      security:
        - BearerAuth: []
      parameters:
        - name: labels
          in: query
          description: Label filter (e.g. role:spine,site:dc1)
          schema:
            type: string
            example: "role:spine"
        - name: status
          in: query
          description: Status filter (ONLINE, OFFLINE, TESTING, DRAINING, MAINTENANCE, DELETED)
          schema:
            type: string
            example: "ONLINE"
        - name: limit
          in: query
          description: Maximum number of targets to return
          schema:
            type: integer
            default: 50
            example: 50
        - name: cursor
          in: query
          description: Pagination cursor
          schema:
            type: string
            example: "eyJpZCI6MTIzfQ=="
      responses:
        '200':
          description: List of targets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    post:
      summary: Register Target
      description: Register a new network target in inventory.
      operationId: createTarget
      tags:
        - Targets
      security:
        - BearerAuth: []
      requestBody:
        description: Target creation payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTargetRequest'
      responses:
        '201':
          description: Target created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Target already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/targets/{id}/drain:
    post:
      summary: Graceful Drain Target
      description: Stop accepting new scenarios for this target and delete after active jobs finish.
      operationId: drainTarget
      tags:
        - Targets
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Target ID or name to drain
          required: true
          schema:
            type: string
            example: "spine1"
      responses:
        '202':
          description: Target entered DRAINING state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrainTargetResponse'
        '404':
          description: Target not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/scenarios:
    get:
      summary: List Scenarios
      description: Retrieve scenarios with orphan/broken status badges.
      operationId: listScenarios
      tags:
        - Scenarios
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Scenario list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioListResponse'

    post:
      summary: Register Scenario
      description: Register YAML scenario DSL (returns warnings if referenced targets are missing).
      operationId: createScenario
      tags:
        - Scenarios
      security:
        - BearerAuth: []
      requestBody:
        description: Scenario definition payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScenarioRequest'
      responses:
        '201':
          description: Scenario registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '400':
          description: DSL Syntax error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/scenarios/{id}/runs:
    post:
      summary: Execute Scenario (Pre-flight & Run)
      description: Run scenario with dynamic inputs, pre-flight target check, and lease lock.
      operationId: runScenario
      tags:
        - Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Scenario identifier
          required: true
          schema:
            type: string
            example: "spine-linkdown-failover"
        - name: Idempotency-Key
          in: header
          description: Unique idempotency key to prevent double runs
          schema:
            type: string
            example: "uuid-12345-67890"
      requestBody:
        description: Dynamic execution inputs
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunJobRequest'
      responses:
        '202':
          description: Scenario execution accepted and started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobExecutionResponse'
        '400':
          description: Invalid input parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Target lock contention
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Pre-flight check failed (target offline/missing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ==========================================
  # Jobs
  # ==========================================
  /v1/jobs:
    get:
      summary: List Job Executions
      description: Retrieve history of scenario execution jobs.
      operationId: listJobs
      tags:
        - Jobs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobDetail'
                  total:
                    type: integer
                    example: 5
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/states/transitions:
    get:
      summary: Query State Transitions
      description: Retrieve state transition diff history timeline.
      operationId: listStateTransitions
      tags:
        - States
      security:
        - BearerAuth: []
      parameters:
        - name: target
          in: query
          description: Target name filter
          schema:
            type: string
            example: "spine1"
        - name: trigger
          in: query
          description: Trigger type filter (inform, polling, scenario_step)
          schema:
            type: string
            example: "inform"
        - name: since
          in: query
          description: ISO timestamp filter
          schema:
            type: string
            example: "2026-08-14T00:00:00Z"
      responses:
        '200':
          description: Transition logs list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateTransitionListResponse'
...
  /v1/audit/logs:
    get:
      summary: Query Audit Logs
      description: Query operation and API audit trails.
      operationId: listAuditLogs
      tags:
        - Audit
      security:
        - BearerAuth: []
      parameters:
        - name: from
          in: query
          description: Start time filter
          schema:
            type: string
            example: "2026-08-01T00:00:00Z"
        - name: to
          in: query
          description: End time filter
          schema:
            type: string
            example: "2026-08-14T23:59:59Z"
      responses:
        '200':
          description: Audit logs list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Target:
      description: Represents a network target device in inventory.
      type: object
      required:
        - id
        - name
        - host
        - port
        - status
        - credential_id
      properties:
        id:
          type: string
          description: Unique target ID
          example: "tgt-spine1-001"
        name:
          type: string
          description: Target unique name
          example: "spine1"
        description:
          type: string
          description: Target description
          example: "Core Spine Switch 01"
        host:
          type: string
          description: IP address or hostname
          example: "192.168.10.1"
        port:
          type: integer
          description: SNMP UDP port
          default: 161
          example: 161
        status:
          type: string
          description: Operational status
          enum: [ONLINE, OFFLINE, TESTING, DRAINING, MAINTENANCE, DELETED]
          example: "ONLINE"
        credential_id:
          type: string
          description: Linked credential profile ID
          example: "cred-v3-admin"
        labels:
          type: object
          description: Key-value metadata labels
          additionalProperties:
            type: string
        polling:
          $ref: '#/components/schemas/TargetPollingConfig'
        created_at:
          type: string
          description: Creation timestamp
          example: "2026-08-14T10:00:00Z"
        updated_at:
          type: string
          description: Last update timestamp
          example: "2026-08-14T10:00:00Z"
    TargetPollingConfig:
      description: Polling configuration profile for target.
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether polling is active
          example: true
        fast_poll_oids:
          type: array
          description: 1s fast poll OID list
          items:
            type: string
            example: "IF-MIB::ifOperStatus.1"
        bulk_get_mibs:
          type: array
          description: 10min bulk get MIB table list
          items:
            type: string
            example: "IF-MIB"
    CreateTargetRequest:
      description: Request payload for registering a new target.
      type: object
      required:
        - name
        - host
        - credential_id
      properties:
        name:
          type: string
          description: Unique name
          example: "spine1"
        description:
          type: string
          description: Description
          example: "Core Spine Switch 01"
        host:
          type: string
          description: IP address or hostname
          example: "192.168.10.1"
        port:
          type: integer
          description: SNMP port
          default: 161
          example: 161
        credential_id:
          type: string
          description: Credential profile ID
          example: "cred-v3-admin"
        labels:
          type: object
          description: Key-value labels
          additionalProperties:
            type: string
        polling:
          $ref: '#/components/schemas/TargetPollingConfig'
...
    TargetListResponse:
      description: Response payload for listing targets.
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          description: Array of target objects
          items:
            $ref: '#/components/schemas/Target'
        total:
          type: integer
          description: Total count
          example: 4
        next_cursor:
          type: string
          description: Next page cursor
          example: "eyJpZCI6NDV9"
...
    DrainTargetResponse:
      description: Response payload after initiating target drain.
      type: object
      required:
        - id
        - status
        - active_jobs_count
      properties:
        id:
          type: string
          description: Target ID
          example: "spine1"
        status:
          type: string
          description: Status
          example: "DRAINING"
        active_jobs_count:
          type: integer
          description: Number of active jobs draining
          example: 1
...
    Scenario:
      description: Scenario summary entity.
      type: object
      required:
        - id
        - name
        - current_version
      properties:
        id:
          type: string
          description: Scenario ID
          example: "spine-linkdown-failover"
        name:
          type: string
          description: Scenario name
          example: "spine-linkdown-failover"
        description:
          type: string
          description: Description
          example: "Spine linkdown and failover verification"
        current_version:
          type: integer
          description: Current version
          example: 1
        is_orphan:
          type: boolean
          description: Whether scenario references missing targets
          example: false
        warnings:
          type: array
          description: Registration warnings
          items:
            type: string
            example: "Target 'spine1' does not exist in inventory"
        created_at:
          type: string
          description: Created timestamp
          example: "2026-08-14T10:00:00Z"
...
    CreateScenarioRequest:
      description: Request payload for registering a new scenario.
      type: object
      required:
        - name
        - dsl_yaml
      properties:
        name:
          type: string
          description: Scenario name
          example: "spine-linkdown-failover"
        description:
          type: string
          description: Description
          example: "Spine linkdown test"
        dsl_yaml:
          type: string
          description: YAML DSL definition
          example: "name: spine-linkdown-failover\nsteps: []"
...
    ScenarioListResponse:
      description: Response payload for listing scenarios.
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          description: Scenarios array
          items:
            $ref: '#/components/schemas/Scenario'
        total:
          type: integer
          description: Total count
          example: 12
...
    RunJobRequest:
      description: Dynamic input variables for scenario run.
      type: object
      properties:
        inputs:
          type: object
          description: Key-value dynamic inputs
          additionalProperties: true
    JobExecutionResponse:
      description: Response payload after accepting scenario execution.
      type: object
      required:
        - job_id
        - status
        - stream_url
      properties:
        job_id:
          type: string
          description: Generated job ID
          example: "job-spine-001"
        status:
          type: string
          description: Initial status
          enum: [QUEUED, RUNNING]
          example: "RUNNING"
        locked_targets:
          type: array
          description: Targets locked by this execution
          items:
            type: string
            example: "spine1"
        stream_url:
          type: string
          description: SSE stream URL to follow progress
          example: "/v1/events/streams?topics=job.step_advanced"
    JobDetail:
      description: Full details of job progress, steps, outputs, and verdicts.
      type: object
      required:
        - id
        - scenario_id
        - status
        - steps
      properties:
        id:
          type: string
          description: Job ID
          example: "job-spine-001"
        scenario_id:
          type: string
          description: Scenario ID
          example: "spine-linkdown-failover"
        status:
          type: string
          description: Job status
          enum: [QUEUED, RUNNING, SUCCESS, FAILED, ABORTED]
          example: "SUCCESS"
        dynamic_inputs:
          type: object
          description: Applied dynamic inputs
          additionalProperties: true
        locked_targets:
          type: array
          description: Locked targets
          items:
            type: string
            example: "spine1"
        steps:
          type: array
          description: Executed steps details
          items:
            $ref: '#/components/schemas/JobStepDetail'
        started_at:
          type: string
          description: Started timestamp
          example: "2026-08-14T10:00:00Z"
        finished_at:
          type: string
          description: Finished timestamp
          example: "2026-08-14T10:00:05Z"
    JobStepDetail:
      description: Details of a single executed scenario step.
      type: object
      required:
        - step_id
        - status
      properties:
        step_id:
          type: string
          description: Step ID
          example: "step1_down"
        status:
          type: string
          description: Step outcome status
          enum: [PENDING, RUNNING, SUCCESS, FAILED, SKIPPED]
          example: "SUCCESS"
        result_output:
          type: object
          description: Step output data
          additionalProperties: true
        error:
          type: string
          description: Error message if failed
          example: ""
        executed_at:
          type: string
          description: Execution timestamp
          example: "2026-08-14T10:00:01Z"
...
    StateTransitionListResponse:
      description: Response payload for listing state transition logs.
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          description: Array of state transition logs
          items:
            $ref: '#/components/schemas/StateTransitionLog'
        total:
          type: integer
          description: Total count
          example: 100
    StateTransitionLog:
      description: State transition log entity recording before/after values.
      type: object
      required:
        - id
        - target
        - state_key
        - old_value
        - new_value
        - trigger
        - created_at
      properties:
        id:
          type: string
          description: Transition log ID
          example: "stl-101"
        target:
          type: string
          description: Target name
          example: "spine1"
        state_key:
          type: string
          description: OID or state key
          example: "raw.spine1.IF-MIB::ifOperStatus.1"
        old_value:
          type: string
          description: Old value before transition
          example: "1"
        new_value:
          type: string
          description: New value after transition
          example: "2"
        trigger:
          type: string
          description: Trigger type
          example: "inform"
        created_at:
          type: string
          description: Timestamp
          example: "2026-08-14T10:00:01.120Z"
...
    AuditLogListResponse:
      description: Response payload for listing operation audit logs.
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          description: Array of audit log entries
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
          description: Total count
          example: 50
...
    ProblemDetails:
      description: RFC 7807 and RFC 9457 Problem Details Standard Error Object.
      type: object
      required:
        - type
        - title
        - status
        - detail
        - code
      properties:
        type:
          type: string
          description: A URI reference identifying the problem type.
          example: "https://musubi.dev/errors/target-in-use"
        title:
          type: string
          description: Short human-readable summary of problem.
          example: "Target In Use"
        status:
          type: integer
          description: HTTP status code.
          example: 409
        detail:
          type: string
          description: Human-readable explanation of this specific occurrence.
          example: "Target 'spine1' is locked by Job #job-001"
        code:
          type: string
          description: Machine-readable stable error code.
          example: "TARGET_IN_USE"
        instance:
          type: string
          description: Request instance URI.
          example: "/v1/targets/spine1"
        actionable_guidance:
          type: object
          description: Resolution guidance and navigation links.
          properties:
            suggestion:
              type: string
              description: Suggested solution
              example: "Wait for active job to finish or use drain mode."
            links:
              type: array
              description: Actionable navigation links
              items:
                $ref: '#/components/schemas/ActionLink'
    ActionLink:
      description: Navigation action link.
      type: object
      properties:
        rel:
          type: string
          description: Relationship
          example: "drain"
        href:
          type: string
          description: Target URL
          example: "/v1/targets/spine1/drain"
        method:
          type: string
          description: HTTP method
          example: "POST"